Generalize base64 command usage in docs
The base64 CLI utility has two different short flag arguments for decode
depending on the platform: -D and -d.

Previously, the docs used the -D flag exclusively with the base64 utility.
Luckily, the long form of the flag is the same across platforms: --decode.

All uses of the base64 -D flag have been replaced with --decode.